Lots of changes
Forgot to commit for a while.

Refactored the routes to display a specific layout based on which page the user is on.

Added proper Firebase integration to fetch users, dogs, and park information. The app will now list off parks that are in Firestore and when a user select them, it will display the users that are currently checked in and also display the dogs info. Need to have ti so when a user clicks on the "more options" part of an attendee it brings up the information on the user and dogs in a modal.

Removed the settings option in the navbar and will eventually place it under "edit profile". I replaced it with a contact page that's not finished, but this will allow users to be able to make suggestions for the app.

Added proper user auth with Firebase. Need to set it up so that when a user creates an account, it will automatically add them to Firestore with their profile pic, along with their dog's info and such.

Added util functions that will fetch their respective information and return a custom object to make it easy for the frontend to work with the data and only pull in what's necessary.

TODO:
1) Make it so that when a user signs up and hasn't put in their dog's info, force them back to that page so in order to use the app they have to finish their profile.
2) Add protected routes.
3) Add contact form for suggestions.
4) Add checkin/out functionality, as well as update Firestore.
5) Add functionality to automatically check someone out after 30 minutes of being checked in.
6) so many other things I can't think of