Better UX
This update improved a lot of UX problems regarding the sign up process. I had to restructure how I created users and documents related to them. The main focus on this one was to make sure the user can only go through the setup process once. That was a pain in the ass figuring that part out. I'm incredibly exhausted and I some how managed to figure it out. Today was a good day.