Testing real time updates for attendees
I refactored Firestore a bit so that real time updates to the park can be feasible. Whenever a user checks in, their document updates with the park and checkin time (this has not changed), but now it also updates the park's document with an array of attendees. This change allows the real time updates to work. I might refactor the "fetchAttendeesAndDogs" function to specifically query the park instead of the users in a specific park. In the short term, either approach is fine, but querying the park instead would definitely scale much better.